refactor: 도메인 패키지 구조 세분화 및 클래스 이동
