+++ src/main/java/domain/ElephantFormation.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.position.Position;
+
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
