+++ src/main/java/domain/InitialPieceLocation.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.position.Position;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
