+++ src/main/java/domain/board/BoardReader.java
@@ -1,6 +1,7 @@
-package domain;
+package domain.board;
 
 import domain.piece.Piece;
+import domain.position.Position;
 
 public interface BoardReader {
