+++ src/main/java/domain/piece/Chariot.java
@@ -1,6 +1,11 @@
 package domain.piece;
 
 import domain.*;
+import domain.board.BoardReader;
+import domain.board.Palace;
+import domain.position.Direction;
+import domain.position.MoveDirection;
+import domain.position.Position;
 
 public class Chariot extends Piece {
